move mq struct
