--- cdc/model/mq.go
@@ -1,99 +1,0 @@
-package model
-
-import (
-	"bytes"
-	"encoding/json"
-
-	"github.com/pingcap/errors"
-	"github.com/pingcap/parser/model"
-)
-
-// MqMessageType is the type of message
-type MqMessageType int
-
-const (
-	// MqMessageTypeUnknow is unknown type of message key
-	MqMessageTypeUnknow MqMessageType = iota
-	// MqMessageTypeRow is row type of message key
-	MqMessageTypeRow
-	// MqMessageTypeDDL is ddl type of message key
-	MqMessageTypeDDL
-	// MqMessageTypeResolved is resolved type of message key
-	MqMessageTypeResolved
-)
-
-const (
-	// BatchVersion1 represents the version of batch format
-	BatchVersion1 uint64 = 1
-)
-
-// MqMessageKey represents the message key
-type MqMessageKey struct {
-	Ts     uint64        `json:"ts"`
-	Schema string        `json:"scm,omitempty"`
-	Table  string        `json:"tbl,omitempty"`
-	Type   MqMessageType `json:"t"`
-}
-
-// Encode encodes the message to the json bytes
-func (m *MqMessageKey) Encode() ([]byte, error) {
-	return json.Marshal(m)
-
-}
-
-// Decode decodes the message from json bytes
-func (m *MqMessageKey) Decode(data []byte) error {
-	return json.Unmarshal(data, m)
-}
-
-// MqMessageRow represents the row message value
-type MqMessageRow struct {
-	Update map[string]*Column `json:"u,omitempty"`
-	Delete map[string]*Column `json:"d,omitempty"`
-}
-
-// Encode encodes the message to the json bytes
-func (m *MqMessageRow) Encode() ([]byte, error) {
-	return json.Marshal(m)
-}
-
-// Decode decodes the message from json bytes
-func (m *MqMessageRow) Decode(data []byte) error {
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.UseNumber()
-	err := decoder.Decode(m)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	for _, column := range m.Update {
-		column.formatVal()
-	}
-	for _, column := range m.Delete {
-		column.formatVal()
-	}
-	return nil
-}
-
-// MqMessageDDL represents the DDL message value
-type MqMessageDDL struct {
-	Query string           `json:"q"`
-	Type  model.ActionType `json:"t"`
-}
-
-// Encode encodes the message to the json bytes
-func (m *MqMessageDDL) Encode() ([]byte, error) {
-	return json.Marshal(m)
-}
-
-// Decode decodes the message from json bytes
-func (m *MqMessageDDL) Decode(data []byte) error {
-	return json.Unmarshal(data, m)
-}
-
-// NewResolvedMessage creates a new message key which of type is Resolved
-func NewResolvedMessage(ts uint64) *MqMessageKey {
-	return &MqMessageKey{
-		Ts:   ts,
-		Type: MqMessageTypeResolved,
-	}
-}
